% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAME.R
\name{rd_fc2_revised}
\alias{rd_fc2_revised}
\title{rd_fc2_revised}
\usage{
rd_fc2_revised(XB, theta, U, d, Y, s2, meaningful_NA)
}
\arguments{
\item{XB}{Time-length list of XB matrix}

\item{theta}{a T x n matrix of nodal effects}

\item{U}{current value of U}

\item{d}{current value of d}

\item{Y}{relational array of relations}

\item{s2}{variance of normal error}

\item{meaningful_NA}{position of structural NA}
}
\value{
One sample estimate of d from normal distribution
}
\description{
Sample d with structural NA
}
