% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAME.R
\name{rtaur_fc}
\alias{rtaur_fc}
\title{rtaur_fc}
\usage{
rtaur_fc(d, a, b)
}
\arguments{
\item{d}{current value of d}

\item{a}{inverse-gamma shape parameter}

\item{b}{inverse-gamma scale parameter}
}
\value{
One sample estimate of tau_u from inverse-gamma distribution
}
\description{
Sample the variance in d
}
