% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAME.R
\name{ru_fc_NA_revised}
\alias{ru_fc_NA_revised}
\title{ru_fc_NA_revised}
\usage{
ru_fc_NA_revised(XB, theta, U, d, Y, cinv, tau_u, s2, meaningful_NA_rows)
}
\arguments{
\item{XB}{Time-length list of XB matrix}

\item{theta}{a T x n matrix of nodal effects}

\item{U}{current value of U}

\item{d}{current value of d}

\item{Y}{relational array of relations}

\item{cinv}{inverse of Gaussian process covariance matrix}

\item{tau_u}{Gaussian process covariance}

\item{s2}{variance of normal errorU}

\item{meaningful_NA_rows}{position of structural NA}
}
\value{
One sample estimate of d from normal distribution
}
\description{
Sample u with structural NA
}
